+++ api/src/main/java/com/grupo4/todolist/api/Domain/Greeting.java
@@ -1,4 +1,4 @@
-package com.grupo4.todolist.api.Model;
+package com.grupo4.todolist.api.Domain;
 
 public record Greeting(long id, String content) {
