+++ api/src/main/java/com/grupo4/todolist/api/Domain/TaskServices.java
@@ -1,17 +1,17 @@
-package com.grupo4.todolist.api.Model;
+package com.grupo4.todolist.api.Domain;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import com.google.gson.Gson;
-import com.grupo4.todolist.api.Model.DAOs.TaskDAO;
-import com.grupo4.todolist.api.Model.Entities.Task;
-import com.grupo4.todolist.api.Model.Repositories.TaskRepo;
+import com.grupo4.todolist.api.Domain.DAOs.TaskDAO;
+import com.grupo4.todolist.api.Domain.Entities.Task;
+import com.grupo4.todolist.api.Domain.Repositories.TaskRepo;
 
-public class TaskModel {
+public class TaskServices {
     private final TaskRepo taskRepo;
 
-    public TaskModel() {
+    public TaskServices() {
         this.taskRepo = new TaskRepo();
     }
